Rename deployment.ts to deploymentConfig.ts to reflect reality
